used consul master determiner to set cron plans

+++ cronhelper.py
@@ -1,52 +1,86 @@
 # -*- coding: utf-8 -*-
 
 from crontab import CronTab
+from threading import Thread
+from time import sleep
+from mysqlhelper import DBConnection
 import pathlib
 import atexit
 
 
-# Use example crontab
-# cron.remove_all()
-# for job in cron:
-#     print(job)
-	# cron.remove(job)
-# job = cron.new(command='touch /home/horifox/SmartHomeApi/test_file_cron')
-# job.minute.every(1)
-# cron.write()
+# Timeout auto-update plans on leader node if necessary
+UPDATE_PLANS_TIMEOUT = 20
 
 
-class CronManager():
-	def __init__(self, api, plan_list, api_config, logger):
+class CronThread():
+
+	submodule_line = '> Cron-submodule: '
+	update_timeout = 20
+	leader_flag = False
+	plans_list = []
+
+	def __init__(self, api, api_config, logger):
 		self.logger = logger
 		self.api_config = api_config
 		self.api = api
 		self.cron = CronTab(user=True)
-		# Очистку нужно проводить до выключения сервера!
+
+
+	def connect_db(self):
+		link_bd = DBConnection(self.logger, user=self.api_config['user_mysql'],
+								password=self.api_config['password_mysql'],
+								host=self.api_config['host_mysql'],
+								port=self.api_config['port_mysql'],
+								database=self.api_config['database_mysql'])
+		return link_bd
+
+
+	def cron_loop(self):
+		self.logger.debug('%sStart cron thread' % self.submodule_line)
+		self.leader_flag = True
+		link_bd = self.connect_db()
+		self.plans_list = link_bd.select('plans', json=True)
 		self.cron.remove_all()
 		self.cron.write()
-		# Load plan when lauch
-		self.load_plan(plan_list)
+		self.load_plan(self.plans_list)
 		atexit.register(self.on_exit)
+		sleep(UPDATE_PLANS_TIMEOUT)
+
+		while self._thread_loop:
+			self.new_plans_list = link_bd.select('plans', json=True)
+			if self.plans_list != self.new_plans_list:
+				self.logger.debug('%sSince the plan data has changed - update cron plans' % self.submodule_line)
+				self.plans_list = self.new_plans_list
+				self.cron.remove_all()
+				self.cron.write()
+				self.load_plan(self.plans_list)
+			sleep(UPDATE_PLANS_TIMEOUT)
 
 
 	def load_plan(self, plan_list):
+		if not self.leader_flag:
+			return
+
 		if len(plan_list) != 0:
 			for elem in plan_list:
 				self.create_plan(str(elem['PlanId']), elem['PlanDays'],
 									elem['PlanTime'], 'relay', elem['ModuleIp'])
-		self.logger.info('List of jobs in cron:')
+		self.logger.info('%sList of jobs in cron:' % self.submodule_line)
 		for job in self.cron:
 			self.logger.info(job)
 		self.logger.info('\n')
 
 
 	def create_plan(self, plan_id, plan_days, plan_time, module_type, module_ip):
+		if not self.leader_flag:
+			return
+
 		# Clear plans to avoid duplication
 		self.delete_plan(plan_id)
 
 		self.logger.info('[!]Load plan:\n')
 		self.logger.info('[P]Id: %s\n[P]Days: %s\n[P]Time: %s\n[P]Type module: %s\n[P]Module ip %s\n' 
-								% (plan_id, plan_days, plan_time, module_type, module_ip))
+								% (plan_id, plan_days, plan_time.replace(' ', '-'), module_type, module_ip))
 
 		week_list = ['ПН', 'ВТ', 'СР', 'ЧТ', 'ПТ', 'СБ', 'ВС']
 		# Convert days to number 0-6
@@ -58,8 +92,9 @@ class CronManager():
 		# Start time
 		id_job_start = plan_id + 'start'
 		job = self.cron.new(comment=id_job_start,
-							command='python3 %s/jobworker.py %s %s %s %s' % 
+							command='python3 %s/cronworker.py %s %s %s %s %s' % 
 													(pathlib.Path(__file__).parent.absolute(),
+													self.api_config['dacrover_host'],
 													self.api_config['dacrover_port'],
 													module_type,
 													module_ip,
@@ -71,8 +106,9 @@ class CronManager():
 			hour_end, minute_end = time_all[1].split(':')
 			id_job_end = plan_id + 'end'
 			job = self.cron.new(comment=id_job_end,
-								command='python3 %s/jobworker.py %s %s %s %s' % 
+								command='python3 %s/cronworker.py %s %s %s %s %s' % 
 														(pathlib.Path(__file__).parent.absolute(),
+														self.api_config['dacrover_host'],
 														self.api_config['dacrover_port'],
 														module_type,
 														module_ip,
@@ -83,6 +119,9 @@ class CronManager():
 
 
 	def delete_plan(self, id_jobs):
+		if not self.leader_flag:
+			return
+
 		id_job_start = id_jobs + 'start'
 		id_job_end = id_jobs + 'end'
 		self.cron.remove_all(comment=id_job_start)
@@ -90,8 +129,20 @@ class CronManager():
 		self.cron.write()
 
 
+	def run(self):
+		self._thread_loop = True
+		thread = Thread(target=self.cron_loop)
+		thread.setDaemon(True)
+		thread.start()
+
+
+	def stop(self):
+		self._thread_loop = False
+		#self.led_thread.join()
+
+	
 	def on_exit(self):
-		self.logger.info('exit handler check stop service')
+		self.logger.info('%sExit handler check stop service' % self.submodule_line)
 		if hasattr(self, 'cron') and self.cron:
 			self.cron.remove_all()
-		self.logger.info('finish exit operation crontab')
+		self.logger.info('%sFinish exit operation crontab' % self.submodule_line)
